# compare_to command: add geometric mean