--- a/doc/cli.rst
+++ b/doc/cli.rst
@@ -112,12 +112,51 @@ two-sample, two-tailed t-test
 <https://en.wikipedia.org/wiki/Student's_t-test>`_ with alpha equals to
 ``0.95``.
 
-Example::
+If the benchmark suites contain more than one benchmark, the `geometric mean
+<https://en.wikipedia.org/wiki/Geometric_mean>`_ of benchmark results compared
+to the reference suite is computed. It is a convenient index to summarize
+benchmark suite results compared to the reference suite. See `How not to lie
+with statistics: the correct way to summarize benchmark results
+<https://www.cse.unsw.edu.au/~cs9242/11/papers/Fleming_Wallace_86.pdf>`_ paper
+by Philip J. Fleming and John J. Wallace (ACM, 1986).
+
+If the geometric mean is greater than 1.0, its benchmark suite is faster than
+the reference.  If it is less than 1.0, its benchmark suite is slower than the
+refrence.
+
+Example 1 comparing Python 3.8 to Python 3.6::
 
     $ python3 -m pyperf compare_to py36.json py38.json
     Mean +- std dev: [py36] 4.70 us +- 0.18 us -> [py38] 4.22 us +- 0.08 us: 1.11x faster (-10%)
 
-On this example, py36 is faster and so used as the reference.
+On this example, py36 is the reference: py38 is faster than py36.
+
+Example 2 comparing two suites (Python 3.7 and Python 3.8) to a reference suite
+(Python 3.6)::
+
+    $ python3 -m pyperf compare_to pyperf/tests/mult_list_py36.json  pyperf/tests/mult_list_py37.json  pyperf/tests/mult_list_py38.json --table
+    +----------------+----------------+------------------------------+------------------------------+
+    | Benchmark      | mult_list_py36 | mult_list_py37               | mult_list_py38               |
+    +================+================+==============================+==============================+
+    | [1]*1000       | 2.13 us        | 2.09 us: 1.02x faster (-2%)  | not significant              |
+    +----------------+----------------+------------------------------+------------------------------+
+    | [1,2]*1000     | 3.70 us        | 5.28 us: 1.42x slower (+42%) | 3.18 us: 1.16x faster (-14%) |
+    +----------------+----------------+------------------------------+------------------------------+
+    | [1,2,3]*1000   | 4.61 us        | 6.05 us: 1.31x slower (+31%) | 4.17 us: 1.11x faster (-10%) |
+    +----------------+----------------+------------------------------+------------------------------+
+    | Geometric mean | (ref)          | 0.82 (slower)                | 1.09 (faster)                |
+    +----------------+----------------+------------------------------+------------------------------+
+
+On this example, mult_list_py36 (Python 3.6) is the reference. According to
+geometric mean, overall mult_list_py37 (Python 3.7) is slower than
+mult_list_py36, whereas overall mult_list_py38 (Python 3.8) is faster than
+mult_list_py36.
+
+The geometric mean is a convenient index to summarize the 3 benchmark results
+of each suite as a single index which is normalized to the reference suite
+results. For example, mult_list_py37 is faster on one benchmark and slower on
+two others: according to the geometric mean, overall it is slower than the
+reference.
 
 See also the ``--compare-to`` :ref:`option of the Runner CLI <runner_cli>`.