--- a/pyperf/tests/mult_list_bench.py
+++ b/pyperf/tests/mult_list_bench.py
@@ -1,0 +1,13 @@
+#!/usr/bin/env python3
+import pyperf
+
+runner = pyperf.Runner()
+runner.timeit("[1]*1000",
+              stmt="[1]*1000",
+              duplicate=1024)
+runner.timeit("[1,2]*1000",
+              stmt="[1,2]*1000",
+              duplicate=1024)
+runner.timeit("[1,2,3]*1000",
+              stmt="[1,2,3]*1000",
+              duplicate=1024)