--- a/pyperf/tests/test_perf_cli.py
+++ b/pyperf/tests/test_perf_cli.py
@@ -7,8 +7,8 @@ import pyperf
 from pyperf import tests
 
 
-TELCO = os.path.join(os.path.dirname(__file__), 'telco.json')
-TRACK_MEMORY = os.path.join(os.path.dirname(__file__), 'track_memory.json')
+TESTDIR = os.path.dirname(__file__)
+TELCO = os.path.join(TESTDIR, 'telco.json')
 
 
 class BaseTestCase(object):
@@ -209,6 +209,79 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
         stdout = self.run_command(*args, **kwargs)
         self.assertEqual(stdout.rstrip(), textwrap.dedent(expected).strip())
 
+    def test_compare_to_geo_mean(self):
+        py36 = os.path.join(TESTDIR, 'mult_list_py36.json')
+        py37 = os.path.join(TESTDIR, 'mult_list_py37.json')
+        py38 = os.path.join(TESTDIR, 'mult_list_py38.json')
+
+        # 2 files
+        expected = """
+            [1]*1000: Mean +- std dev: [mult_list_py36] 2.13 us +- 0.06 us -> [mult_list_py37] 2.09 us +- 0.04 us: 1.02x faster (-2%)
+            [1,2]*1000: Mean +- std dev: [mult_list_py36] 3.70 us +- 0.05 us -> [mult_list_py37] 5.28 us +- 0.09 us: 1.42x slower (+42%)
+            [1,2,3]*1000: Mean +- std dev: [mult_list_py36] 4.61 us +- 0.13 us -> [mult_list_py37] 6.05 us +- 0.11 us: 1.31x slower (+31%)
+
+            Geometric mean: 0.82 (slower)
+        """
+        self.check_command(expected, 'compare_to', py36, py37)
+
+        # 2 files (grouped by speed)
+        expected = """
+            Slower (2):
+            - [1,2]*1000: 3.70 us +- 0.05 us -> 5.28 us +- 0.09 us: 1.42x slower (+42%)
+            - [1,2,3]*1000: 4.61 us +- 0.13 us -> 6.05 us +- 0.11 us: 1.31x slower (+31%)
+
+            Faster (1):
+            - [1]*1000: 2.13 us +- 0.06 us -> 2.09 us +- 0.04 us: 1.02x faster (-2%)
+
+            Geometric mean: 0.82 (slower)
+        """
+        self.check_command(expected, 'compare_to', "--group-by-speed", py36, py37)
+
+        # 3 files
+        expected = """
+            [1]*1000
+            ========
+
+            Mean +- std dev: [mult_list_py36] 2.13 us +- 0.06 us -> [mult_list_py37] 2.09 us +- 0.04 us: 1.02x faster (-2%)
+            Mean +- std dev: [mult_list_py36] 2.13 us +- 0.06 us -> [mult_list_py38] 2.13 us +- 0.03 us: 1.00x slower (+0%)
+            Not significant!
+
+            [1,2]*1000
+            ==========
+
+            Mean +- std dev: [mult_list_py36] 3.70 us +- 0.05 us -> [mult_list_py37] 5.28 us +- 0.09 us: 1.42x slower (+42%)
+            Mean +- std dev: [mult_list_py36] 3.70 us +- 0.05 us -> [mult_list_py38] 3.18 us +- 0.08 us: 1.16x faster (-14%)
+
+            [1,2,3]*1000
+            ============
+
+            Mean +- std dev: [mult_list_py36] 4.61 us +- 0.13 us -> [mult_list_py37] 6.05 us +- 0.11 us: 1.31x slower (+31%)
+            Mean +- std dev: [mult_list_py36] 4.61 us +- 0.13 us -> [mult_list_py38] 4.17 us +- 0.11 us: 1.11x faster (-10%)
+
+            Geometric mean
+            ==============
+
+            mult_list_py37: 0.82 (slower)
+            mult_list_py38: 1.09 (faster)
+        """
+        self.check_command(expected, 'compare_to', py36, py37, py38)
+
+        # 3 files as table
+        expected = """
+            +----------------+----------------+------------------------------+------------------------------+
+            | Benchmark      | mult_list_py36 | mult_list_py37               | mult_list_py38               |
+            +================+================+==============================+==============================+
+            | [1]*1000       | 2.13 us        | 2.09 us: 1.02x faster (-2%)  | not significant              |
+            +----------------+----------------+------------------------------+------------------------------+
+            | [1,2]*1000     | 3.70 us        | 5.28 us: 1.42x slower (+42%) | 3.18 us: 1.16x faster (-14%) |
+            +----------------+----------------+------------------------------+------------------------------+
+            | [1,2,3]*1000   | 4.61 us        | 6.05 us: 1.31x slower (+31%) | 4.17 us: 1.11x faster (-10%) |
+            +----------------+----------------+------------------------------+------------------------------+
+            | Geometric mean | (ref)          | 0.82 (slower)                | 1.09 (faster)                |
+            +----------------+----------------+------------------------------+------------------------------+
+        """
+        self.check_command(expected, 'compare_to', '--table', py36, py37, py38)
+
     def test_hist(self):
         # Force terminal size on Python 3 for shutil.get_terminal_size()
         env = dict(os.environ)
@@ -325,7 +398,8 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             Run 4: 0 warmups, 1 value, 2^15 loops
             - value 1: 7208.0 kB
         """
-        stdout = self.run_command('dump', TRACK_MEMORY)
+        track_memory = os.path.join(TESTDIR, 'track_memory.json')
+        stdout = self.run_command('dump', track_memory)
         self.assertIn(textwrap.dedent(expected).strip(), stdout)
 
     def test_dump_quiet(self):