--- a/pyperf/tests/test_perf_cli.py
+++ b/pyperf/tests/test_perf_cli.py
@@ -220,7 +220,7 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             [1,2]*1000: Mean +- std dev: [mult_list_py36] 3.70 us +- 0.05 us -> [mult_list_py37] 5.28 us +- 0.09 us: 1.42x slower (+42%)
             [1,2,3]*1000: Mean +- std dev: [mult_list_py36] 4.61 us +- 0.13 us -> [mult_list_py37] 6.05 us +- 0.11 us: 1.31x slower (+31%)
 
-            Geometric mean: 0.82 (slower)
+            Geometric mean: 1.22 (slower)
         """
         self.check_command(expected, 'compare_to', py36, py37)
 
@@ -233,7 +233,7 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             Faster (1):
             - [1]*1000: 2.13 us +- 0.06 us -> 2.09 us +- 0.04 us: 1.02x faster (-2%)
 
-            Geometric mean: 0.82 (slower)
+            Geometric mean: 1.22 (slower)
         """
         self.check_command(expected, 'compare_to', "--group-by-speed", py36, py37)
 
@@ -261,8 +261,8 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             Geometric mean
             ==============
 
-            mult_list_py37: 0.82 (slower)
-            mult_list_py38: 1.09 (faster)
+            mult_list_py37: 1.22 (slower)
+            mult_list_py38: 0.92 (faster)
         """
         self.check_command(expected, 'compare_to', py36, py37, py38)
 
@@ -277,7 +277,7 @@ class TestPerfCLI(BaseTestCase, unittest.TestCase):
             +----------------+----------------+------------------------------+------------------------------+
             | [1,2,3]*1000   | 4.61 us        | 6.05 us: 1.31x slower (+31%) | 4.17 us: 1.11x faster (-10%) |
             +----------------+----------------+------------------------------+------------------------------+
-            | Geometric mean | (ref)          | 0.82 (slower)                | 1.09 (faster)                |
+            | Geometric mean | (ref)          | 1.22 (slower)                | 0.92 (faster)                |
             +----------------+----------------+------------------------------+------------------------------+
         """
         self.check_command(expected, 'compare_to', '--table', py36, py37, py38)