--- a/setup.py
+++ b/setup.py
@@ -26,7 +26,7 @@
 #  - git commit -a -m "post-release"
 #  - git push
 
-VERSION = '2.0.1'
+VERSION = '2.1.0'
 
 DESCRIPTION = 'Python module to run and analyze benchmarks'
 CLASSIFIERS = [