When comparing suites with more than one benchmark, compute the
geometric mean of benchmark speeds, to compare a whole suite to the
reference suite with a single number.