@vstinner 2020-10-13
@methane @pablogsal @serhiy-storchaka: Would you mind to have a look at this new feature? I'm not sure that I compute the geometric mean of the correct thing.

IMO it's a big feature and it should help a lot to compare two benchmark suites using a single number rather than 30 numbers.

A geometric mean has no unit. It's a unusual value. In short, geo mean > 1.0 means "faster", geo mean < 1.0 means "slower".

---

PyPy uses the geometric mean to compare PyPy to CPython as a single number. speed.pypy.org announces:

"The **geometric average of all benchmarks is 0.24** or 4.2 times faster than cpython"

I'm not sure that speed.pypy.org and my PR compute the geometric mean of the same thing, since it announces that 0.24 means "faster".

My PR computes the geometric mean of all "speeds". The speed of a benchmark is the ratio: (benchmark mean) / (reference benchmark mean).