@vstinner 2020-10-13
> I'm not sure that speed.pypy.org and my PR compute the geometric mean of the same thing, since it announces that 0.24 means "faster".

Oops, I computed the geometric mean backwards. I normalized reference / benchmark, but the correct fomula is benchmark / reference.

I completed my PR to update the documentation which explains how the geometric mean is computed and what it means.