@vstinner 2020-10-13
I wrote this PR when I saw  bench_results.txt  of https://bugs.python.org/issue41972 : it's really hard to read all these numbers and understand if it's faster or slower.