@vstinner 2020-10-26
I rewrote this PR as a serie of commits. It's now implemented in the master branch, I close this PR.